requirements.txt
discord.py
python-dotenv
holidays
pytest